requirements.txt
alembic==1.3.2
aniso8601==8.0.0
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
Werkzeug==0.16.0
WTForms==2.2.1
